pyproject.toml
[project]
name = "local-ai-hackathon"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "numpy>=1.26.4",
    "openai-whisper>=20231117",
    "datasets>=2.18.0",
    "torch>=2.2.1",
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "setuptools>=69.2.0",
    "torchaudio>=2.2.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/local_ai_hackathon"]
